requirements.txt
pyautogen
openai
pypdf
chromadb
whisper-mic
pyttsx3
IPython
sentence_transformers